Move ActorExecutionTrace to replay package
